+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # v1.7.0 (2023-09-22)
-- **BREAKING** The MSRV of the library is `1.63.0` stable.
+- **BREAKING** The MSRV of the library is `1.64.0` stable.
 - **BREAKING** The MSRV of the binary is `1.70.0` stable.
 - dependencies updated